+++ handler/src/main/java/io/netty/handler/ssl/OpenSslServerSessionContext.java
@@ -23,7 +23,7 @@ import org.apache.tomcat.jni.SSLContext;
  * {@link OpenSslSessionContext} implementation which offers extra methods which are only useful for the server-side.
  */
 public final class OpenSslServerSessionContext extends OpenSslSessionContext {
-    OpenSslServerSessionContext(OpenSslContext context) {
+    OpenSslServerSessionContext(ReferenceCountedOpenSslContext context) {
         super(context);
     }
