+++ handler/src/main/java/io/netty/handler/ssl/OpenSslSessionContext.java
@@ -32,13 +32,13 @@ public abstract class OpenSslSessionContext implements SSLSessionContext {
     private static final Enumeration<byte[]> EMPTY = new EmptyEnumeration();
 
     private final OpenSslSessionStats stats;
-    final OpenSslContext context;
+    final ReferenceCountedOpenSslContext context;
 
     // IMPORTANT: We take the OpenSslContext and not just the long (which points the native instance) to prevent
     //            the GC to collect OpenSslContext as this would also free the pointer and so could result in a
     //            segfault when the user calls any of the methods here that try to pass the pointer down to the native
     //            level.
-    OpenSslSessionContext(OpenSslContext context) {
+    OpenSslSessionContext(ReferenceCountedOpenSslContext context) {
         this.context = context;
         stats = new OpenSslSessionStats(context);
     }
