+++ handler/src/test/java/io/netty/handler/ssl/OpenSslEngineTest.java
@@ -21,6 +21,7 @@ import io.netty.handler.ssl.ApplicationProtocolConfig.SelectedListenerFailureBeh
 import io.netty.handler.ssl.ApplicationProtocolConfig.SelectorFailureBehavior;
 import io.netty.handler.ssl.util.InsecureTrustManagerFactory;
 import io.netty.handler.ssl.util.SelfSignedCertificate;
+import io.netty.util.ReferenceCountUtil;
 import io.netty.util.internal.ThreadLocalRandom;
 import org.junit.BeforeClass;
 import org.junit.Test;
@@ -79,26 +80,33 @@ public class OpenSslEngineTest extends SSLEngineTest {
     }
     @Test
     public void testWrapHeapBuffersNoWritePendingError() throws Exception {
-        final SslContext clientContext = SslContextBuilder.forClient()
+        clientSslCtx = SslContextBuilder.forClient()
                 .trustManager(InsecureTrustManagerFactory.INSTANCE)
                 .sslProvider(sslProvider())
                 .build();
         SelfSignedCertificate ssc = new SelfSignedCertificate();
-        SslContext serverContext = SslContextBuilder.forServer(ssc.certificate(), ssc.privateKey())
+        serverSslCtx = SslContextBuilder.forServer(ssc.certificate(), ssc.privateKey())
                 .sslProvider(sslProvider())
                 .build();
-        SSLEngine clientEngine = clientContext.newEngine(UnpooledByteBufAllocator.DEFAULT);
-        SSLEngine serverEngine = serverContext.newEngine(UnpooledByteBufAllocator.DEFAULT);
-        handshake(clientEngine, serverEngine);
+        SSLEngine clientEngine = null;
+        SSLEngine serverEngine = null;
+        try {
+            clientEngine = clientSslCtx.newEngine(UnpooledByteBufAllocator.DEFAULT);
+            serverEngine = serverSslCtx.newEngine(UnpooledByteBufAllocator.DEFAULT);
+            handshake(clientEngine, serverEngine);
 
-        ByteBuffer src = ByteBuffer.allocate(1024 * 10);
-        ThreadLocalRandom.current().nextBytes(src.array());
-        ByteBuffer dst = ByteBuffer.allocate(1);
-        // Try to wrap multiple times so we are more likely to hit the issue.
-        for (int i = 0; i < 100; i++) {
-            src.position(0);
-            dst.position(0);
-            assertSame(SSLEngineResult.Status.BUFFER_OVERFLOW, clientEngine.wrap(src, dst).getStatus());
+            ByteBuffer src = ByteBuffer.allocate(1024 * 10);
+            ThreadLocalRandom.current().nextBytes(src.array());
+            ByteBuffer dst = ByteBuffer.allocate(1);
+            // Try to wrap multiple times so we are more likely to hit the issue.
+            for (int i = 0; i < 100; i++) {
+                src.position(0);
+                dst.position(0);
+                assertSame(SSLEngineResult.Status.BUFFER_OVERFLOW, clientEngine.wrap(src, dst).getStatus());
+            }
+        } finally {
+            cleanupSslEngine(clientEngine);
+            cleanupSslEngine(serverEngine);
         }
     }
